requirements/development.txt
-r base.txt

django-kombu==0.9.2
